alura-spring-boot/api/src/main/java/med/voll/api/domain/consulta/validadores/datas/ValidaFinalSemana.java
package med.voll.api.domain.consulta.validadores.datas;

import jakarta.validation.constraints.Future;
import jakarta.validation.constraints.NotNull;
import med.voll.api.domain.ValidacaoException;
import med.voll.api.domain.consulta.DadosAgendamentoConsulta;
import med.voll.api.domain.consulta.validadores.ValidadorAgendamentoDeConsulta;
import org.springframework.stereotype.Component;

import java.time.DayOfWeek;
import java.time.LocalDateTime;
@Component
public class ValidaFinalSemana implements ValidadorAgendamentoDeConsulta {

    public void validar(DadosAgendamentoConsulta dados) {
        DayOfWeek dayOfWeek = dados.data().getDayOfWeek();
        if (dayOfWeek == DayOfWeek.SATURDAY || dayOfWeek == DayOfWeek.SUNDAY) {
            throw new ValidacaoException("Data fornecida está fora do dia de funcionamento da clinica!");
        }

    }

}
